+++ configure.py
@@ -761,7 +761,7 @@ scylla_core = (['replica/database.cc',
                 'cql3/maps.cc',
                 'cql3/values.cc',
                 'cql3/expr/expression.cc',
-                'cql3/expr/to_restriction.cc',
+                'cql3/expr/restrictions.cc',
                 'cql3/expr/prepare_expr.cc',
                 'cql3/functions/user_function.cc',
                 'cql3/functions/functions.cc',
